refactor(core): move dictionary utilities
