+++ orchestrator/core/legacy/migrators/discoveryspace/entitysource_to_samplestore.py
@@ -4,12 +4,12 @@
 """Legacy migrator for renaming entitySourceIdentifier to sampleStoreIdentifier"""
 
 from orchestrator.core.legacy.registry import legacy_migrator
-from orchestrator.core.legacy.utils import (
+from orchestrator.core.resources import CoreResourceKinds
+from orchestrator.utilities.dictionaries import (
     get_nested_value,
     remove_nested_field,
     set_nested_value,
 )
-from orchestrator.core.resources import CoreResourceKinds
 
 
 @legacy_migrator(
