+++ orchestrator/core/legacy/migrators/samplestore/entitysource_migrations.py
@@ -4,12 +4,12 @@
 """Legacy validators for migrating entitysource to samplestore naming"""
 
 from orchestrator.core.legacy.registry import legacy_migrator
-from orchestrator.core.legacy.utils import (
+from orchestrator.core.resources import CoreResourceKinds
+from orchestrator.utilities.dictionaries import (
     get_nested_value,
     has_nested_field,
     set_nested_value,
 )
-from orchestrator.core.resources import CoreResourceKinds
 
 
 @legacy_migrator(
@@ -280,7 +280,7 @@ def remove_specification_storage_location(data: dict) -> dict:
         return data
 
     # Remove config.specification.storageLocation if it exists
-    from orchestrator.core.legacy.utils import remove_nested_field
+    from orchestrator.utilities.dictionaries import remove_nested_field
 
     remove_nested_field(data, "config.specification.storageLocation")
